mcp-server/requirements.txt
mcp[cli]
requests
langchain 
langchain-mcp-adapters 
python-dotenv
langgraph
langchain-groq